+++ src/main/java/com/ail/revolut/app/logic/AccountManager.java
@@ -1,6 +1,6 @@
 package com.ail.revolut.app.logic;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import com.ail.revolut.app.exception.NotEnoughFundsException;
 
@@ -8,9 +8,9 @@ public interface AccountManager {
 
 	Long createAccount();
 
-	BigInteger getBalance(Long id);
+	BigDecimal getBalance(Long id);
 
-	void deposit(Long id, BigInteger amount);
+	void deposit(Long id, BigDecimal amount);
 
-	void withdraw(Long id, BigInteger amount) throws NotEnoughFundsException;
+	void withdraw(Long id, BigDecimal amount) throws NotEnoughFundsException;
 }
