+++ src/main/java/com/ail/revolut/app/logic/TransferManager.java
@@ -1,10 +1,10 @@
 package com.ail.revolut.app.logic;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import com.ail.revolut.app.exception.NotEnoughFundsException;
 
 public interface TransferManager {
 
-	void transfer(Long fromAccountId, Long toAccountId, BigInteger amount) throws NotEnoughFundsException;
+	void transfer(Long fromAccountId, Long toAccountId, BigDecimal amount) throws NotEnoughFundsException;
 }
